feat: add poduct routes and local file upload

+++ internal/utils/response.go
@@ -19,10 +19,10 @@ type PaginatedResponse struct {
 }
 
 type PaginatedMeta struct {
-	Total      int `json:"total"`
-	Page       int `json:"page"`
-	PageSize   int `json:"page_size"`
-	TotalPages int `json:"total_pages"`
+	Total      int64 `json:"total"`
+	Page       int   `json:"page"`
+	PageSize   int   `json:"page_size"`
+	TotalPages int   `json:"total_pages"`
 }
 
 func SuccessResponse(c *gin.Context, message string, data any) {
